+++ package.json
@@ -28,7 +28,9 @@
     "zustand": "^4.4.1"
   },
   "devDependencies": {
+    "autoprefixer": "^10.4.16",
     "eslint-config-prettier": "^9.0.0",
+    "postcss": "^8.4.30",
     "prettier": "^2.8.8",
     "sass": "^1.66.1",
     "tailwindcss": "^3.3.3"
