+++ src/app/detail/page.jsx
@@ -1,0 +1,13 @@
+'use client'
+import UserDetail from "@/components/userDetail/UserDetail";
+
+const Detail = () => {
+    return (
+        <>
+            <h1>Detial</h1>
+            <UserDetail/>
+        </>
+    )
+}
+
+export default Detail;
